Add separate annotations for identity and authserver configs
