+++ src/main/java/org/radarbase/management/config/annotations/AuthServerEnabled.kt
@@ -1,0 +1,15 @@
+package org.radarbase.management.config.annotations
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
+import org.springframework.context.annotation.Configuration
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+@Configuration
+@ConditionalOnProperty(
+    name = ["managementportal.authServer.internal"],
+    havingValue = "true",
+    matchIfMissing = true
+)
+annotation class AuthServerEnabled
+
